v.1
This is a semi-completed version where drawing title was being pulled in addition to drawing numbers and page numbers. In the next version this function will be removed, but may be added later if proper pdf points are able to be identified and LTText boxes can be identified and pulled for that data.